Create Purchases Component
Component very similar to Cards one. Save the User's payments through the AddPayment component, list every user's payments next to this one in a table. Everytime the user delete or add a new payment the table must be automatically updated with a new page render.

Every Payment Object is composed of:
- Product Purchased Name
- Date of Purchase
- Cost of the Product
- Category o Tag of the Product
- Card ID used (the user will be shown the Card Name  in the AddPayment component)