Fix Methods on Card Component
Now the money parameter is casted as int before to be registered in the database, this bring to zero any error of the function isDisabled() on AddComponent to check if the cost is minus than the money on the card.

Now everytime a card that has a purchase is deleted, the purchased reference to the card is set to false.