Fix Bug on AddPurchase Component
Add propriety "name" to Date Textfield. Before this, any date the user used to choose were not picked by the function onChange() and this cause the registration of this actual day as a date in the db. Now that this propriety has been set the user can choose any date.