Add Expenses Properties in Card Object
Now every card tracks its expenses. Once a card is created this voice is set to 0. As the user register a payment this value add the cost of the Payment. In this way with every card it is possibile to know how many money have been spent. If a purchase is deleted, now the card restores the cost of the purchase to the money value, and update the expenses voice taking off the value of each payment.